Use geometry structs for layout and Skia interop
Replace separate width/height/margin/page fields with PtSize, PtOffset,
PtRect, PtLineSegment and PtEdgeInsets. Add From impls to convert Pt
types to skia_safe types. Update model, parsing, layout, rendering and
tests to use the new geometry types and adjust font API to accept Pt.

+++ src/render/layout/measurer.rs
@@ -40,26 +40,14 @@ impl TextMeasurer {
         bold: bool,
         italic: bool,
     ) -> Pt {
-        let font = fonts::make_font(
-            &self.font_mgr,
-            font_family,
-            f32::from(font_size),
-            bold,
-            italic,
-        );
+        let font = fonts::make_font(&self.font_mgr, font_family, font_size, bold, italic);
         let (width, _) = font.measure_str(text, None);
         Pt::new(width)
     }
 
     /// Get the line height (ascent + descent + leading) for a font.
     pub fn line_height(&self, font_family: &str, font_size: Pt, bold: bool, italic: bool) -> Pt {
-        let font = fonts::make_font(
-            &self.font_mgr,
-            font_family,
-            f32::from(font_size),
-            bold,
-            italic,
-        );
+        let font = fonts::make_font(&self.font_mgr, font_family, font_size, bold, italic);
         let (_, metrics) = font.metrics();
         Pt::new(-metrics.ascent + metrics.descent + metrics.leading)
     }
